refactor(lib): extract "podcast::TaskPool" to own module
outside of "::podcast" for general use
