+++ tui/src/ui/model/mod.rs
@@ -45,8 +45,9 @@ use std::sync::mpsc::{self, Receiver, Sender};
 use std::time::{Duration, Instant};
 use termusiclib::config::{ServerOverlay, SharedServerSettings, SharedTuiSettings};
 use termusiclib::library_db::TrackDB;
-use termusiclib::podcast::{db::Database as DBPod, Podcast, PodcastFeed, TaskPool};
+use termusiclib::podcast::{db::Database as DBPod, Podcast, PodcastFeed};
 use termusiclib::songtag::SongTag;
+use termusiclib::taskpool::TaskPool;
 use termusiclib::utils::get_app_config_path;
 use termusicplayback::{PlayerCmd, Playlist};
 use tokio::sync::mpsc::UnboundedSender;
